Remove explicit use of `long`
Removes use of `long` for PY3-compatibility, and fixes some other
PY3-issues in the affected scripts and modules.

Issue: CRB-4082

+++ testsuite/tests/test_core/test_core_Entity/test_EntitySpread.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """ Basic tests for Cerebrum.Entity.EntitySpread. """
 import pytest
@@ -140,7 +139,7 @@ def test_list_all_with_spread(entity_obj, entities):
 
 
 def test_list_entity_spreads(entity_obj, entities, entity_type):
-    expected = [(long(ent['entity_id']), long(int(spread)))
+    expected = [(int(ent['entity_id']), int(spread))
                 for ent in entities
                 for spread in ent['spreads']]
     entity_types = {ent['entity_type'] for ent in entities}
